Merge the `client` module into the root module
The entire project is intended to "just" be a client, so it didn't
particularly make sense to have a separate module for that.
Merging in the module makes the overall code structure simpler, and also
fixes another issue: before this, if the project was cloned from the
repository, it actually wouldn't build! There was previously an empty
`world` module at the root, which was never used.
This has been replaced with the `client::world` module, and so no empty
modules are left.